requirements.txt
Appium-Python-Client==0.26
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
ddt==1.2.1
easygui==0.96
enum34==1.1.6
et-xmlfile==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
lxml==4.2.1
more-itertools==4.1.0
mysql-connector-python==8.0.15
namedlist==1.7
numpy==1.16.1
openpyxl==2.6.0
pandas==0.24.1
parsel==1.5.1
pluggy==0.6.0
protobuf==3.7.0
py==1.5.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pygal==1.7.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytest==3.5.0
pytest-allure-adaptor==1.7.10
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
selenium==3.12.0
service-identity==18.1.0
six==1.11.0
urllib3==1.24.1
w3lib==1.19.0
xlrd==1.2.0
zope.interface==4.6.0
